requirements.txt
supervision
ultralytics
numpy
roboflow
PyYAML